refactor(backup): create @suite-common/backup with shared backup logic
