+++ suite-common/backup/src/index.ts
@@ -1,0 +1,5 @@
+export * from './backupUtils';
+export * from './createAdditionalBackup';
+export * from './shamirConstants';
+export * from './shamirUtils';
+export type * from './types';
