+++ suite-common/backup/src/types.ts
@@ -1,0 +1,5 @@
+export type AdditionalBackupPhase = 'verify-ownership' | 'backup';
+
+export type AdditionalBackupResult =
+    | { success: true }
+    | { success: false; phase: AdditionalBackupPhase };
