+++ suite/backup/src/backupUtils.ts
@@ -1,4 +1,4 @@
-import { type PROTO } from '@trezor/connect';
+export { hasNonWordlistBackup, isBackupComplete } from '@suite-common/backup';
 
 import { type ConfirmKey } from './types';
 
@@ -9,9 +9,3 @@ export const canContinue = (userConfirmed: ConfirmKey[], isDeviceLocked?: boolea
     (['wrote-seed-properly', 'made-no-digital-copy', 'will-hide-seed'] as const).every(e =>
         userConfirmed.includes(e),
     ) && !isDeviceLocked;
-
-export const isBackupComplete = (features: PROTO.Features): boolean =>
-    features.backup_availability === 'NotAvailable';
-
-export const hasNonWordlistBackup = (features: PROTO.Features): boolean =>
-    features.backup_type != null && features.backup_type !== 'Bip39';
