+++ suite/recovery/src/isRecoveryInProgress.ts
@@ -1,9 +1,6 @@
 import { type PROTO } from '@trezor/connect';
 
-export const isAdditionalShamirBackupInProgress = (features: PROTO.Features) =>
-    features.recovery_status === 'Backup' &&
-    (features.recovery_type === undefined || features.recovery_type === null) &&
-    features.backup_availability === 'Available';
+export { isAdditionalShamirBackupInProgress } from '@suite-common/backup';
 
 export const isRecoveryInProgress = (features: PROTO.Features) => {
     const { recovery_status, backup_availability } = features;
