+++ suite/recovery/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        { "path": "../../suite-common/backup" },
         { "path": "../../suite-common/device" },
         {
             "path": "../../suite-common/redux-utils"
